pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "iati-sphinx-theme"
readme = "README.md"
dynamic = ["version", "description"]
dependencies = []

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "Flake8-pyproject",
    "mypy",
    "sphinx-autobuild",
    "types-docutils"
]

[project.entry-points."sphinx.html_themes"]
iati_sphinx_theme = "iati_sphinx_theme"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88

[tool.mypy]
strict = true
